# public/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Life@USTC API",
    "version": "1.0.0",
    "description": "OpenAPI document generated from Next.js routes by next-openapi-gen"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/admin/comments": {
      "get": {
        "operationId": "get-api-admin-comments",
        "summary": "List moderation comments.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["active", "softbanned", "deleted"]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/adminCommentsResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users": {
      "get": {
        "operationId": "get-api-admin-users",
        "summary": "List users for admin console.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/adminUsersResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar-subscriptions/{id}": {
      "get": {
        "operationId": "get-api-calendar-subscriptions-{id}",
        "summary": "Get subscription detail.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {},
            "required": true,
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/calendarSubscriptionSchema"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch-api-calendar-subscriptions-{id}",
        "summary": "Update sections in one subscription.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {},
            "required": true,
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/calendarSubscriptionUpdateRequestSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/calendarSubscriptionSummarySchema"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-api-calendar-subscriptions-{id}",
        "summary": "Delete one subscription.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {},
            "required": true,
            "example": "123"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/admin/comments/{id}": {
      "patch": {
        "operationId": "patch-api-admin-comments-{id}",
        "summary": "Moderate one comment.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/adminModerateCommentRequestSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/adminModeratedCommentResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "adminModerateCommentRequestSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["active", "softbanned", "deleted"]
          },
          "moderationNote": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["status"]
      },
      "adminModeratedCommentResponseSchema": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/components/schemas/adminCommentBaseSchema"
          }
        },
        "required": ["comment"]
      },
...
      "adminCommentsResponseSchema": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/adminCommentSchema"
            }
          }
        },
        "required": ["comments"]
      },
...
      "adminUsersResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/adminUserListItemSchema"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/pageMetaSchema"
          }
        },
        "required": ["data", "pagination"]
      },
...
      "calendarSubscriptionSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "course": {
                  "$ref": "#/components/schemas/courseBaseSchema"
                },
                "semester": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/semesterSchema"
                    }
                  ]
                },
                "campus": {
                  "type": "object"
                },
                "schedules": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/scheduleWithRoomTeachersSchema"
                  }
                }
              },
              "required": ["course", "semester", "campus", "schedules"]
            }
          }
        },
        "required": ["id", "userId", "sections"]
      },
      "calendarSubscriptionUpdateRequestSchema": {
        "type": "object",
        "properties": {
          "sectionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true
            }
          }
        },
        "required": ["sectionIds"]
      },
      "calendarSubscriptionSummarySchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "sections": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer"
                }
              },
              "required": ["id"]
            }
          }
        },
        "required": ["id", "sections"]
      },
...
      "semesterSchema": {
        "type": "object",
        "properties": {
          "startDate": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          },
          "endDate": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          }
        },
        "required": ["startDate", "endDate"]
      },
...
      "dateTimeSchema": {
        "type": "string"
      },
      "pageMetaSchema": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": ["page", "pageSize", "total", "totalPages"]
      },
...
      "courseBaseSchema": {
        "type": "object"
      },
...
      "teacherSchema": {
        "type": "object"
      },
...
      "roomTypeSchema": {
        "type": "object"
      },
      "roomWithBuildingSchema": {
        "type": "object",
        "properties": {
          "building": {
            "type": "object"
          },
          "roomType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/roomTypeSchema"
              }
            ]
          }
        },
        "required": ["building", "roomType"]
      },
      "scheduleWithRoomTeachersSchema": {
        "type": "object",
        "properties": {
          "date": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          },
          "room": {
            "allOf": [
              {
                "$ref": "#/components/schemas/roomWithBuildingSchema"
              }
            ]
          },
          "teachers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/teacherSchema"
            }
          }
        },
        "required": ["date", "room", "teachers"]
      },
      "adminUserListItemSchema": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/dateTimeSchema"
          }
        },
        "required": ["email", "createdAt"]
      },
...
      "adminCommentBaseSchema": {
        "type": "object",
        "properties": {
          "createdAt": {
            "$ref": "#/components/schemas/dateTimeSchema"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/dateTimeSchema"
          },
          "deletedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          },
          "moderatedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          }
        },
        "required": ["createdAt", "updatedAt", "deletedAt", "moderatedAt"]
      },
...
      "adminCommentSchema": {
        "type": "object",
        "properties": {
          "createdAt": {
            "$ref": "#/components/schemas/dateTimeSchema"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/dateTimeSchema"
          },
          "deletedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          },
          "moderatedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          },
          "user": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "nullable": true
              }
            },
            "required": ["name"],
            "nullable": true
          },
          "section": {
            "type": "object",
            "properties": {
              "jwId": {
                "type": "integer"
              },
              "code": {
                "type": "string"
              }
            },
            "required": ["jwId", "code"],
            "nullable": true
          },
          "course": {
            "type": "object",
            "properties": {
              "jwId": {
                "type": "integer"
              },
              "code": {
                "type": "string"
              },
              "nameCn": {
                "type": "string"
              }
            },
            "required": ["jwId", "code", "nameCn"],
            "nullable": true
          },
          "teacher": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "nameCn": {
                "type": "string"
              }
            },
            "required": ["id", "nameCn"],
            "nullable": true
          },
          "homework": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "section": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  }
                },
                "required": ["code"]
              }
            },
            "required": ["id", "title", "section"],
            "nullable": true
          },
          "sectionTeacher": {
            "type": "object",
            "properties": {
              "section": {
                "type": "object",
                "properties": {
                  "jwId": {
                    "type": "integer"
                  },
                  "code": {
                    "type": "string"
                  }
                },
                "required": ["jwId", "code"]
              },
              "teacher": {
                "type": "object",
                "properties": {
                  "nameCn": {
                    "type": "string"
                  }
                },
                "required": ["nameCn"]
              }
            },
            "required": ["section", "teacher"],
            "nullable": true
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "deletedAt",
          "moderatedAt",
          "user",
          "section",
          "course",
          "teacher",
          "homework",
          "sectionTeacher"
        ]
      },
...
    },
...
  }
}